tests/ambient-light/w3c/AmbientLightSensor_secure_context.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: secure context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<h2>Precondition</h2>
<ol>
  <li>
    This is an insecurity context.
  </li>
</ol>
<script>

async_test(function(t) {
  try {
    let sensor = new AmbientLightSensor();
    sensor.onchange = t.step_func_done(function(event) {
      assert_not_equals(String(sensor.reading), "[object AmbientLightSensorReading]");
      sensor.stop();
    });
    sensor.start();
  } catch (e) {
    assert_equals(e.name, "SecurityError");
  }
}, "Test that sensor readings must only be available within a secure context");

</script>
